requirements/dev.txt
-r base.txt

django-debug-toolbar
django-extensions
jinja2
pyzmq
tornado